adding custom grafity to the script
todo for next time:

there should be 2 vectors. One for the player, and one for the camera. and the requested movement should be derived from the camera and applied to the player.
Aka, right should be the rigt of the camera, which is just force applied to the body. The model will follow the direction.
And the camera will try to follow the model.
then you have full circle again